fix: Resolve MyPy type errors
- Cast numpy scalar types to float in tiered_storage_strategy.py
- Convert Redis keys() result to list before unpacking in engine.py
- Add type annotation for tier_counts dict in server.py

+++ src/core/standards/tiered_storage_strategy.py
@@ -334,9 +334,9 @@ class TieredStorageStrategy:
         if access_counts:
             import numpy as np
             stats.update({
-                "access_count_mean": np.mean(access_counts),
-                "access_count_median": np.median(access_counts),
-                "access_count_p95": np.percentile(access_counts, 95),
+                "access_count_mean": float(np.mean(access_counts)),
+                "access_count_median": float(np.median(access_counts)),
+                "access_count_p95": float(np.percentile(access_counts, 95)),
                 "documents_above_threshold": sum(
                     1 for c in access_counts if c >= self.access_threshold
                 )
